MAILET-115 Remove abstract getReplyTo implementation
