+++ server/mailet/mailets/src/main/java/org/apache/james/transport/mailets/DSNBounce.java
@@ -146,6 +146,11 @@ public class DSNBounce extends AbstractRedirect {
     }
 
     @Override
+    protected MailAddress getReplyTo() throws MessagingException {
+        return SpecialAddress.NULL;
+    }
+
+    @Override
     protected MailAddress getReversePath(Mail originalMail) {
         return SpecialAddress.NULL;
     }
