+++ server/mailet/mailets/src/test/java/org/apache/james/transport/mailets/redirect/AbstractRedirectTest.java
@@ -74,6 +74,11 @@ public class AbstractRedirectTest {
         protected InternetAddress[] getTo() throws MessagingException {
             return null;
         }
+
+        @Override
+        protected MailAddress getReplyTo() throws MessagingException {
+            return null;
+        }
     }
 
     @Test
